requirements.txt
openai==1.68.2
mysql-connector-python==9.2.0
simplejson==3.20.1
faiss-cpu==1.10.0
python-dotenv==1.0.1
Jinja2==3.1.6
langchain-openai==0.3.14
langchain-community==0.3.21
langchain-upstage==0.6.0
pydantic-settings==2.8.1
pytest==8.3.5
black==23.11.0
flake8==6.1.0
isort==6.0.1
mypy==1.15.0
